Remove unused types and functions no longer needed post-pydantic-ai migration
