--- src/sqlsaber/models/__init__.py
@@ -1,10 +1,0 @@
-"""Models module for SQLSaber."""
-
-from .types import ColumnInfo, ForeignKeyInfo, SchemaInfo, ToolDefinition
-
-__all__ = [
-    "ColumnInfo",
-    "ForeignKeyInfo",
-    "SchemaInfo",
-    "ToolDefinition",
-]
